Add ListBase in favour of MessageListBase
